2.0.0.2: diagnostics: Fixed float/double missmatch.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@821 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

Former-commit-id: ec9b1c41ffcefa6da96a13733632d60b75981747

+++ common/diagnostics/graph.cpp
@@ -168,7 +168,7 @@ class line : public drawable
 	boost::optional<diagnostics::guide> guide_;
 	boost::circular_buffer<std::pair<double, bool>> line_data_;
 
-	std::vector<float>		tick_data_;
+	std::vector<double>		tick_data_;
 	bool					tick_tag_;
 	color c_;
 public:
